GH-116 Update Groups implementation (#128)
* GH-116 Update Groups implementation

This is to match the latest Bitbucket behaviour, specifically the `auto_add` has been replaced by an additional permission entry `none`. Appears as though groups are auto-added to all repositories with a default permission, that can then be overridden on a per-repository basis.

* GH-116 Address issue with updating Branch Restrictions

Bitbucket's API returns HTTP 400 when trying to update an existing branch restriction, so instead we'll delete and re-create.